backend/requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.20.0,<0.30.0
pymongo>=4.5.0,<5.0.0
minio>=7.1.0,<8.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0
cryptography>=40.0.0,<42.0.0